+++ engine/src/main/java/com/github/emilienkia/oraclaestus/types/FloatType.java
@@ -1,6 +1,6 @@
 package com.github.emilienkia.oraclaestus.types;
 
-public class FloatType extends NumberType<Float> {
+public class FloatType extends NumberType<Double> {
 
     static final FloatType INSTANCE = new FloatType();
 
@@ -12,26 +12,26 @@ public class FloatType extends NumberType<Float> {
     }
 
     @Override
-    public Float cast(Object value) {
+    public Double cast(Object value) {
         switch(value) {
             case null:
                 return null;
             case Float f:
-                return f;
+                return f.doubleValue();
             case Double d:
-                return d.floatValue();
+                return d;
             case Long l:
-                return l.floatValue();
+                return l.doubleValue();
             case Integer i:
-                return i.floatValue();
+                return i.doubleValue();
             case String str:
                 try {
-                    return Float.parseFloat(str);
+                    return Double.parseDouble(str);
                 } catch (NumberFormatException e) {
                     throw new IllegalArgumentException("Cannot convert String to Float: " + value, e);
                 }
             case Boolean b:
-                return b ? 1.0f : 0.0f; // Convert Boolean to Float (true -> 1.0, false -> 0.0)
+                return b ? 1.0 : 0.0; // Convert Boolean to Float (true -> 1.0, false -> 0.0)
             default:
                 throw new IllegalArgumentException("Cannot convert " + value.getClass().getName() + " to Float");
         }
